INTRD-16012 - [Account-operation] - matching 2 AO with amount less than 1 does not work

+++ opencell-admin/ejbs/src/main/java/org/meveo/service/payments/impl/MatchingCodeService.java
@@ -226,7 +226,7 @@ public class MatchingCodeService extends PersistenceService<MatchingCode> {
                 }
             }
 
-            if(0 != amountToMatch.longValue()){
+            if(0 != amountToMatch.doubleValue()){
                 accountOperation.setMatchingAmount(accountOperation.getMatchingAmount().add(amountToMatch));
             accountOperation.setUnMatchingAmount(accountOperation.getUnMatchingAmount().subtract(amountToMatch));
             accountOperation.setMatchingStatus(fullMatch ? MatchingStatusEnum.L : MatchingStatusEnum.P);
@@ -312,7 +312,7 @@ public class MatchingCodeService extends PersistenceService<MatchingCode> {
                 
             }
 
-            if(0 != amountToMatch.longValue()) {
+            if(0 != amountToMatch.doubleValue()) {
                 accountOperation.setMatchingAmount(accountOperation.getMatchingAmount().add(amountToMatch));
                 accountOperation.setUnMatchingAmount(accountOperation.getUnMatchingAmount().subtract(amountToMatch));
                 accountOperation.setMatchingStatus(fullMatch ? MatchingStatusEnum.L : MatchingStatusEnum.P);
